pyproject.toml
[project]
name = "fastapiproject3"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "bcrypt>=4.3.0",
    "black>=25.1.0",
    "fastapi>=0.115.11",
    "pydantic-settings>=2.8.1",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "schemas>=0.7.1",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
]
